--- desk/src/components/units/auth/logout/Logout.container.tsx
@@ -1,5 +1,0 @@
-import LogoutUI from "./Logout.presenter";
-
-export default function Logout() {
-	return <LogoutUI />
-}
